requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
black==23.1.0
blobconverter==1.3.0
boto3==1.26.50
botocore==1.29.50
bs4==0.0.1
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==40.0.1
dataclasses-json==0.5.7
depthai==2.19.1.0
depthai-sdk==1.9.1.1
distinctipy==1.2.2
distlib==0.3.6
docker==6.0.1
dotenv-python==0.0.1
dulwich==0.20.50
ffmpy3==0.2.4
filelock==3.8.2
frozenlist==1.3.3
html5lib==1.1
idna==3.4
importlib-metadata==5.2.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
langchain @ git+https://github.com/ItsJustLukas78/langchain@ab6461f632a9bc4501c02785b6291f9eaad9a88f
lockfile==0.12.2
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.17.0
marshmallow-enum==1.5.1
mcap==0.3.0
mcap-ros1-support==0.0.8
more-itertools==9.0.0
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.24.1
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.5.5.62
packaging==22.0
pathspec==0.11.1
pexpect==4.8.0
pkginfo==1.9.2
platformdirs==2.6.0
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.14.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyrsistent==0.19.2
python-dateutil==2.8.2
pytube==12.1.2
PyTurboJPEG==1.6.4
pyusb==1.2.1
PyYAML==6.0
Qt.py==1.3.7
rapidfuzz==2.13.7
requests==2.26.0
requests-toolbelt==0.10.1
rosbags==0.9.11
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sentry-sdk==1.5.1
shellingham==1.5.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
sympy==1.12
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.6
torch==2.0.1
trove-classifiers==2022.12.22
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
webencodings==0.5.1
webget==1.0
websocket-client==1.4.2
xattr==0.10.1
xmltodict==0.13.0
yarl==1.9.2
zipp==3.11.0
zstandard==0.19.0
